Raise Eq36Exception from eq36 submodule
It is common for EQ3/6 to fail with some kind of error. We want to be
able to handle these errors in the sailors, for example. When an error
occurs, EQ3/6 write the error messages to standard output. We parse that
output for error messages, wrap the first in an Eq36Exception, and
raise.