Implement a parameter registry to disambiguate parameters
When the Boatswain attempts to apply constraints, it identifies
parameters based on value (in the pythonic sense). This results in
randomly chosen values being assigned to multiple parameters in the
navigator.

We could have used the builtin `id` method, but that's a little more
risky if we start adding parallelism inside of the navigator, for
example.

Instead, when we construct the Boatswain, we register all of the
parameters it is considering to get a unique integer for each. I don't
know how reentrant this is, but it works for now.